+++ options.go
@@ -39,6 +39,23 @@ func WithLinkTransformer(f LinkTransformerFunc) testoplugin.Option {
 	}
 }
 
+// WithTimeouts enables test timeout handling.
+// When enabled, allure plugin will mark all
+// running tests and steps as broken just a few milliseconds
+// before [`t.Deadline`] and write their reports.
+//
+// By default, it's enabled.
+//
+// [`t.Deadline`]: https://pkg.go.dev/github.com/ozontech/testo#T.Deadline
+func WithTimeouts(enable bool) testoplugin.Option {
+	return testoplugin.Option{
+		Propagate: true,
+		Value: option(func(a *PluginAllure) {
+			a.handleTimeouts = enable
+		}),
+	}
+}
+
 // WithGroupHooks configures BeforeAll & AfterAll hooks groupping
 // in the allure report.
 //